changed stream implementations to use 64-bit integer as size

+++ inc/TCCodec.h
@@ -30,7 +30,7 @@
 // License along with this library; if not, write to the Free Software       
 // Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA.
 //*******************************************************************************
-//  $Id: TCCodec.h 957 2010-01-28 23:17:00Z the_____tiger $
+//  $Id$
 //*******************************************************************************
 #ifndef _TCCODEC_H_
 #define _TCCODEC_H_
@@ -70,91 +70,91 @@ namespace TC
         * @param val the short which should store the short
         * @return Number of bytes decoded
         */
-       virtual uint32 Decode(Stream& stream, sint16 &val) = 0;
+       virtual uint64 Decode(Stream& stream, sint16 &val) = 0;
        /**
         * Method for reading one unsigned 16bit integer from the stream
         * @param stream Stream from which to decode the data
         * @param val the short which should store the short
         * @return Number of bytes decoded
         */
-       virtual uint32 Decode(Stream& stream, uint16 &val) = 0;
+       virtual uint64 Decode(Stream& stream, uint16 &val) = 0;
        /**
         * Method for reading one signed int 32bit integer the stream
         * @param stream Stream from which to decode the data
         * @param val the int which should store the int
         * @return Number of bytes decoded
         */
-       virtual uint32 Decode(Stream& stream, sint32 &val) = 0;
+       virtual uint64 Decode(Stream& stream, sint32 &val) = 0;
        /**
         * Method for reading one unsigned 32bit integer from the stream
         * @param stream Stream from which to decode the data
         * @param val the int which should store the int
         * @return Number of bytes decoded
         */
-       virtual uint32 Decode(Stream& stream, uint32 &val) = 0;
+       virtual uint64 Decode(Stream& stream, uint32 &val) = 0;
        /**
         * Method for reading one signed 64bit integer from the stream
         * @param stream Stream from which to decode the data
         * @param val the long which should store the long
         * @return Number of bytes decoded
         */
-       virtual uint32 Decode(Stream& stream, sint64 &val) = 0;
+       virtual uint64 Decode(Stream& stream, sint64 &val) = 0;
        /**
         * Method for reading one unsigned 64bit integer from the stream
         * @param stream Stream from which to decode the data
         * @param val the long which should store the long
         * @return Number of bytes decoded
         */
-       virtual uint32 Decode(Stream& stream, uint64 &val) = 0;
+       virtual uint64 Decode(Stream& stream, uint64 &val) = 0;
        /**
         * Method for reading one float from the stream
         * @param stream Stream from which to decode the data
         * @param val the float which should store the float
         * @return Number of bytes decoded
         */
-       virtual uint32 Decode(Stream& stream, float &val) = 0;
+       virtual uint64 Decode(Stream& stream, float &val) = 0;
        /**
         * Method for reading one double from the stream
         * @param stream Stream from which to decode the data
         * @param val the double which should store the double
         * @return Number of bytes decoded
         */
-       virtual uint32 Decode(Stream& stream, double &val) = 0;
+       virtual uint64 Decode(Stream& stream, double &val) = 0;
        /**
         * Method for reading one string from the stream
         * @param stream Stream from which to decode the data
         * @param val the string which should store the string
         * @return Number of bytes decoded
         */
-       virtual uint32 Decode(Stream& stream, char *val) = 0;
+       virtual uint64 Decode(Stream& stream, char *val) = 0;
        /**
        * Method for reading one string from the stream
        * @param stream Stream from which to decode the data
        * @param val the string which should store the string
        * @return Number of bytes decoded
        */
-       virtual uint32 Decode(Stream& stream, std::string& val) = 0;
+       virtual uint64 Decode(Stream& stream, std::string& val) = 0;
        /**
        * Method for reading one wstring from the stream
        * @param stream Stream from which to decode the data
        * @param val the string which should store the string
        * @return Number of bytes decoded
        */
-       virtual uint32 Decode(Stream& stream, std::wstring& val) = 0;
+       virtual uint64 Decode(Stream& stream, std::wstring& val) = 0;
        /**
         * Method for reading one char from the stream
         * @param stream Stream from which to decode the data
         * @param val the char which should store the char
         * @return Number of bytes decoded
         */
-       virtual uint32 Decode(Stream& stream, char &val) = 0;
+       virtual uint64 Decode(Stream& stream, char &val) = 0;
        /**
         * Method for reading one unsigned char from the stream
         * @param stream Stream from which to decode the data
         * @param val the char which should store the char
         * @return Number of bytes decoded
         */
-       virtual uint32 Decode(Stream& stream, uchar &val)  = 0;
+       virtual uint64 Decode(Stream& stream, uchar &val)  = 0;
 
 
        /**
@@ -163,105 +163,105 @@ namespace TC
         * @param val the short which to write
         * @return Number of bytes encoded
         */
-       virtual uint32 Encode(sint16 val, Stream& stream) = 0;
+       virtual uint64 Encode(sint16 val, Stream& stream) = 0;
        /**
         * Method for writing one unsigned 16bit integer to the stream
         * @param stream Stream to which to encode the data
         * @param val the short which to write
         * @return Number of bytes encoded
         */
-       virtual uint32 Encode(uint16 val, Stream& stream) = 0;
+       virtual uint64 Encode(uint16 val, Stream& stream) = 0;
        /**
         * Method for writing one signed 32bit integer to the stream
         * @param stream Stream to which to encode the data
         * @param val the int which to write
         * @return Number of bytes encoded
         */
-       virtual uint32 Encode(sint32 val, Stream& stream) = 0;
+       virtual uint64 Encode(sint32 val, Stream& stream) = 0;
        /**
         * Method for writing one unsigned 32bit integer to the stream
         * @param stream Stream to which to encode the data
         * @param val the int which to write
         * @return Number of bytes encoded
         */
-       virtual uint32 Encode(uint32 val, Stream& stream) = 0;
+       virtual uint64 Encode(uint32 val, Stream& stream) = 0;
        /**
         * Method for writing one signed 64bit integer to the stream
         * @param stream Stream to which to encode the data
         * @param val the long which to write
         * @return Number of bytes encoded
         */
-       virtual uint32 Encode(sint64 val, Stream& stream) = 0;
+       virtual uint64 Encode(sint64 val, Stream& stream) = 0;
        /**
         * Method for writing one unsigned 64bit integer to the stream
         * @param stream Stream to which to encode the data
         * @param val the long which to write
         * @return Number of bytes encoded
         */
-       virtual uint32 Encode(uint64 val, Stream& stream) = 0;
+       virtual uint64 Encode(uint64 val, Stream& stream) = 0;
        /**
         * Method for writing one float to the stream
         * @param stream Stream to which to encode the data
         * @param val the float which to write
         * @return Number of bytes encoded
         */
-       virtual uint32 Encode(float val, Stream& stream) = 0;
+       virtual uint64 Encode(float val, Stream& stream) = 0;
        /**
         * Method for writing one double to the stream
         * @param stream Stream to which to encode the data
         * @param val the double which to write
         * @return Number of bytes encoded
         */
-       virtual uint32 Encode(double val, Stream& stream) = 0;
+       virtual uint64 Encode(double val, Stream& stream) = 0;
        /**
        * Method for writing one string to the stream
        * @param stream Stream to which to encode the data
        * @param val the string which to write
        * @return Number of bytes encoded
        */
-       virtual uint32 Encode(const char *val, Stream& stream) = 0;
+       virtual uint64 Encode(const char *val, Stream& stream) = 0;
        /**
         * Method for writing one string to the stream
         * @param stream Stream to which to encode the data
         * @param val the string which to write
         * @return Number of bytes encoded
         */
-       virtual uint32 Encode(const std::string& val, Stream& stream) = 0;
+       virtual uint64 Encode(const std::string& val, Stream& stream) = 0;
        /**
         * Method for writing one wstring to the stream
         * @param stream Stream to which to encode the data
         * @param val the string which to write
         * @return Number of bytes encoded
         */
-       virtual uint32 Encode(const std::wstring& val, Stream& stream) = 0;
+       virtual uint64 Encode(const std::wstring& val, Stream& stream) = 0;
        /**
         * Method for writing one signed char to the stream
         * @param stream Stream to which to encode the data
         * @param val the char which to write
         * @return Number of bytes encoded
         */
-       virtual uint32 Encode(char val, Stream& stream) = 0;
+       virtual uint64 Encode(char val, Stream& stream) = 0;
        /**
         * Method for writing one unsigned char to the stream
         * @param stream Stream to which to encode the data
         * @param val the char which to write
         * @return Number of bytes encoded
         */
-       virtual uint32 Encode(uchar val, Stream& stream) = 0;
+       virtual uint64 Encode(uchar val, Stream& stream) = 0;
 
        /**
         * Writes the end of line characters if needed for the stream
         * @param stream Stream to which to encode the end of line
         * @return Number of bytes encoded
         */
-       virtual uint32 EncodeEndOfLine(Stream& stream) = 0;
+       virtual uint64 EncodeEndOfLine(Stream& stream) = 0;
        /**
         * Writes space to the stream if needed for the stream
         * needed for separating words or numbers
         * @param stream Stream to which to encode the space
         * @return Number of bytes encoded
         */
-       virtual uint32 EncodeSpace(Stream& stream) = 0;
+       virtual uint64 EncodeSpace(Stream& stream) = 0;
 
        virtual ~Codec() {}
     };
